indexer easy: remove error_logs
