+++ crates/sui-indexer/src/lib.rs
@@ -20,17 +20,13 @@ use errors::IndexerError;
 use mysten_metrics::spawn_monitored_task;
 use sui_core::event_handler::EventHandler;
 use sui_json_rpc::{JsonRpcServerBuilder, ServerHandle, CLIENT_SDK_TYPE_HEADER};
-use sui_json_rpc_types::SuiTransactionResponseOptions;
-use sui_sdk::apis::ReadApi as SuiReadApi;
 use sui_sdk::{SuiClient, SuiClientBuilder};
-use sui_types::base_types::TransactionDigest;
 
 use crate::apis::{
     CoinReadApi, EventReadApi, GovernanceReadApi, ReadApi, TransactionBuilderApi, WriteApi,
 };
 use crate::handlers::checkpoint_handler::CheckpointHandler;
 use crate::store::IndexerStore;
-use crate::types::SuiTransactionFullResponse;
 
 pub mod apis;
 pub mod errors;
@@ -191,39 +187,4 @@ pub async fn build_json_rpc_server<S: IndexerStore + Sync + Send + 'static + Clo
         config.rpc_server_port,
     );
     Ok(builder.start(default_socket_addr).await?)
-}
-
-pub async fn multi_get_full_transactions(
-    read_api: &SuiReadApi,
-    digests: Vec<TransactionDigest>,
-) -> Result<Vec<SuiTransactionFullResponse>, IndexerError> {
-    let sui_transactions = read_api
-        .multi_get_transactions_with_options(
-            digests.clone(),
-            // MUSTFIX(gegaowp): avoid double fetching both input and raw_input
-            SuiTransactionResponseOptions::new()
-                .with_input()
-                .with_effects()
-                .with_events()
-                .with_raw_input(),
-        )
-        .await
-        .map_err(|e| {
-            IndexerError::FullNodeReadingError(format!(
-                "Failed to get transactions {:?} with error: {:?}",
-                digests.clone(),
-                e
-            ))
-        })?;
-    let sui_full_transactions: Vec<SuiTransactionFullResponse> = sui_transactions
-        .into_iter()
-        .map(SuiTransactionFullResponse::try_from)
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(|e| {
-            IndexerError::FullNodeReadingError(format!(
-                "Unexpected None value in SuiTransactionFullResponse of digests {:?} with error {:?}",
-                digests, e
-            ))
-        })?;
-    Ok(sui_full_transactions)
 }
